refactor: relocate redact.ts to util/ (carve-out before Experience Engine removal)
